Create annotated chromosome map

Render an interactive graphics visualization of entire chromosomes or chromosomal regions. Annotate multiple targeted regions to visualize probe targets.

chromosome_map(genome, probes, title = "", colours = list(), ...)

Arguments

genome
A tibble indicating the starting and ending position of each chromosome. Contains three columns:

Name of the chromosome

The starting position of the chromosome

The ending position of the chromosome.

probes
A tibble indicating the starting and ending position of each probe. Contains five columns:

A unique identifier for the particular probe

Name of the chromosome the probe is on

The starting position of the probe

The ending position of the probe

An identifier indicating the probe set the probe belongs to.

title
The title of the plot.

colours
A list of colours indicating the annotation colour for each probe set.

...
<dynamic-dots> Additional arguments passed to chromoMap::chromoMap().

See also

Examples

probes <- tibble::tribble(
  ~rowid, ~chrom, ~start, ~end, ~probe_set,
  4487L, "chr14", 2342135L, 2342284L, "IBC",
  2813L, "chr3", 830503L, 830769L, "DR2",
  4673L, "chr5", 482233L, 482391L, "IBC",
  3337L, "chr9", 375274L, 375417L, "IBC",
  2449L, "chr12", 532032L, 532281L, "DR2",
  1565L, "chr7", 383447L, 383653L, "HAP",
  3115L, "chr14", 1401991L, 1402160L, "IBC",
  1446L, "chr4", 734737L, 734936L, "HAP",
  4555L, "chr10", 93054L, 93223L, "IBC",
  3627L, "chr7", 162127L, 162277L, "IBC"
)

chromosome_map(genome_Pf3D7, probes)

chromosome_map(
  genome_Pf3D7,
  probes,
  title = "Example Chromosome Map",
  colours = list(c("#006A8EFF", "#A8A6A7FF", "#B1283AFF"))
)