Label mutations

For each data point, label a mutation as "ref", "alt", "ins", or "del".

Data points that have more reference (REF) than alternate (ALT) calls will be labeled as "ref".

Mutations whose REF and ALT calls are both one base pair and who have more ALT calls than REF calls will be labeled as "alt".

Mutations whose REF and ALT calls differ in length and have more ALT calls than REF calls will be labeled as "ins".

Mutations whose REF and ALT calls differ in length and have more REF calls than ALT calls will be labeled as "del".

Usage

label_mutations(.data, .before = NULL, .after = NULL)

# S3 method for ref_alt_cov_tbl
label_mutations(.data, .before = NULL, .after = NULL)

Arguments

.data

The data set containing REF and ALT calls.

.before, .after

<tidy-select> Optionally, control where new columns should appear (the default is to add to the right-hand side). See dplyr::relocate() for more details.

Value

The object .data with an added column that indicates the mutation type of each row. The column is added to the right-hand side by default, but this may be controlled by the .before and .after arguments.

Examples

# Read example data
data <- read_tbl_ref_alt_cov(
  miplicorn_example("reference_AA_table.csv"),
  miplicorn_example("alternate_AA_table.csv"),
  miplicorn_example("coverage_AA_table.csv"),
  gene == "atp6"
)

# Add ref and alt calls to data
sequences <- c("A", "T", "C", "G", "AT", "TC", "TGC")
data <- dplyr::mutate(
  data,
  ref = sample(sequences, size = nrow(data), replace = TRUE),
  alt = sample(sequences, size = nrow(data), replace = TRUE)
)

# Label the mutations
label_mutations(data)
#> # A ref alt cov table: 260 × 13
#>    sample     gene_id       gene  mutation_name  exonic_func  aa_change targeted
#>    <chr>      <chr>         <chr> <chr>          <chr>        <chr>     <chr>   
#>  1 D10-JJJ-23 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  2 D10-JJJ-43 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  3 D10-JJJ-55 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  4 D10-JJJ-5  PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  5 D10-JJJ-47 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  6 D10-JJJ-15 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  7 D10-JJJ-27 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  8 D10-JJJ-10 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  9 D10-JJJ-28 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#> 10 D10-JJJ-52 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#> # … with 250 more rows, and 6 more variables: ref_umi_count <dbl>,
#> #   alt_umi_count <dbl>, coverage <dbl>, ref <chr>, alt <chr>,
#> #   mutation_label <chr>
label_mutations(data, .after = alt)
#> # A ref alt cov table: 260 × 13
#>    sample     gene_id       gene  mutation_name  exonic_func  aa_change targeted
#>    <chr>      <chr>         <chr> <chr>          <chr>        <chr>     <chr>   
#>  1 D10-JJJ-23 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  2 D10-JJJ-43 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  3 D10-JJJ-55 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  4 D10-JJJ-5  PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  5 D10-JJJ-47 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  6 D10-JJJ-15 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  7 D10-JJJ-27 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  8 D10-JJJ-10 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#>  9 D10-JJJ-28 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#> 10 D10-JJJ-52 PF3D7_0106300 atp6  atp6-Ala623Glu missense_va… Ala623Glu Yes     
#> # … with 250 more rows, and 6 more variables: ref_umi_count <dbl>,
#> #   alt_umi_count <dbl>, coverage <dbl>, ref <chr>, alt <chr>,
#> #   mutation_label <chr>